docs(button api docs): button api docs
