+++ .vscode/settings.json
@@ -16,6 +16,9 @@
   "[javascript]": {
     "editor.defaultFormatter": "esbenp.prettier-vscode"
   },
+  "[markdown]": {
+    "editor.formatOnSave": false,
+  },
   "editor.codeActionsOnSave": {
     "source.fixAll.eslint": true
   }
