+++ src/button/_example/icon.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import { Button } from 'tdesign-mobile-react';
+
+export default function () {
+  return (
+    <>
+      <Button>图标按钮</Button>
+    </>
+  );
+}
