+++ src/button/_example/shape.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import { Button } from 'tdesign-mobile-react';
+
+export default function () {
+  return (
+    <>
+      <Button shape="round">圆角按钮</Button>
+      <Button shape="square">直角按钮</Button>
+    </>
+  );
+}
