+++ src/button/_example/size.jsx
@@ -1,0 +1,17 @@
+import React from 'react';
+import { Button } from 'tdesign-mobile-react';
+
+export default function () {
+  return (
+    <>
+      <Button theme="primary" size="large">
+        按钮 44
+      </Button>
+      <Button theme="primary">按钮 40</Button>
+      <br />
+      <Button theme="primary" size="small">
+        按钮 36
+      </Button>
+    </>
+  );
+}
