+++ src/button/_example/test.jsx
@@ -4,9 +4,17 @@ import { Button } from 'tdesign-mobile-react';
 export default function () {
   return (
     <>
-      <Button>默认按钮</Button>
-      <Button theme="primary">重要按钮</Button>
-      <Button theme="danger">警告按钮</Button>
+      {/* 正确 */}
+      <div>
+        <Button theme="primary">主要按钮</Button>
+        <Button variant="outline">主要按钮</Button>
+      </div>
+
+      {/* 错误 */}
+      <div>
+        <Button theme="primary">主要按钮</Button>
+        <Button theme="primary">主要按钮</Button>
+      </div>
     </>
   );
 }
